Reorg passes tests
