+++ brain/admin/tests/all_data_tests.py
@@ -7,18 +7,6 @@ from models import RepExercisesHistory, RepExercisesTaxonomy
 
 
 class AllDataTests(unittest.TestCase):
-    def test_prepare_history_entry_stringifies_date(self):
-        test_entry = RepExercisesHistory(
-            user_id=1,
-            exercise_id=12,
-            sets=1,
-            reps=10,
-            weight=20,
-            date=date.today()
-        )
-        test_entry = AllData._prepare_history_entry(test_entry)
-        self.assertIsInstance(test_entry.date, str)
-
     def test_prepare_taxonomy_entry_debooleanizes(self):
         exercise_name = 'test_awesomeness_exercise'
         test_entry = RepExercisesTaxonomy(
